src/components: move DrawerMenu
