+++ src/components/global/DrawerMenu.vue
@@ -24,7 +24,7 @@
 import { defineComponent } from 'vue';
 
 // mocks
-import { menuBottom, menuTop } from '../mocks/layout';
+import { menuBottom, menuTop } from '../../mocks/layout';
 
 export default defineComponent({
   name: 'DrawerMenu',
